REFACTOR: chat page
