+++ src/app/(home)/home.module.scss
@@ -1,0 +1,10 @@
+.homePage {
+  display: flex;
+  height: 100%;
+  width: 100vw;
+
+  .main {
+    flex: 3;
+    padding: 24px;
+  }
+}
