+++ src/component/chatMessage/chatMessage.module.scss
@@ -1,0 +1,39 @@
+.chatMessage {
+  display: flex;
+  gap: 12px;
+  align-items: center;
+
+  .bubble {
+    position: relative;
+    background: rgba(128, 128, 128, 0.229);
+    border-radius: 4px;
+    padding: 4px 12px;
+
+    &:after {
+      content: "";
+      position: absolute;
+      left: 0;
+      top: 50%;
+      width: 0;
+      height: 0;
+      border: 10px solid transparent;
+      border-right-color: rgba(128, 128, 128, 0.229);
+      border-left: 0;
+      border-bottom: 0;
+      margin-top: -5px;
+      margin-left: -10px;
+    }
+  }
+}
+
+.chatMessage_isMe {
+  justify-content: end;
+
+  img {
+    display: none;
+  }
+
+  .bubble:after {
+    border-right-color: transparent;
+  }
+}
