+++ src/component/profile/Profile.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import styles from "./chat.module.scss";
+import styles from "./profile.module.scss";
 
 interface ProfileCardProps {
   profileSrc: string;
@@ -9,7 +9,13 @@ interface ProfileCardProps {
 const Profile = ({ profileSrc, nickname }: ProfileCardProps) => {
   return (
     <div className={styles.profile}>
-      <Image className={styles.profile_image} width={24} height={24} src={profileSrc} alt="profile-image" />
+      <Image
+        className={styles.profile_image}
+        width={24}
+        height={24}
+        src={profileSrc || "/profile_image_default.jpeg"}
+        alt="profile-image"
+      />
       <span>{nickname}</span>
     </div>
   );
