+++ src/component/profile/profile.module.scss
@@ -1,0 +1,8 @@
+.profile {
+  display: flex;
+  gap: 8px;
+
+  &_image {
+    border-radius: 50%;
+  }
+}
