--- src/component/sideBar/chat.module.scss
@@ -1,124 +1,0 @@
-.chatPage {
-  display: flex;
-  height: 100%;
-  width: 100vw;
-}
-
-.sidebar {
-  flex: 1;
-  border-right: 1px solid rgba(128, 128, 128, 0.305);
-
-  &_header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 12px;
-    border-bottom: 1px solid rgba(128, 128, 128, 0.305);
-    height: 48px;
-  }
-
-  &_list {
-    padding: 12px;
-  }
-
-  &_profileCard {
-    border: 1px solid rgba(128, 128, 128, 0.305);
-    padding: 8px;
-    border-radius: 4px;
-  }
-}
-
-.chatroom {
-  flex: 3;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-
-  &_header {
-    position: absolute;
-    display: flex;
-    padding: 12px;
-    gap: 8px;
-    background: white;
-    width: 100%;
-    border-bottom: 1px solid rgba(128, 128, 128, 0.305);
-    height: 48px;
-    z-index: 2;
-  }
-
-  &_messages {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    gap: 12px;
-    padding: 12px;
-  }
-
-  &_form {
-    padding: 12px;
-    display: flex;
-    gap: 8px;
-
-    input {
-      flex: 1;
-      padding: 8px 12px;
-      border-radius: 4px;
-      border: 1px solid rgba(128, 128, 128, 0.305);
-    }
-
-    button {
-      width: 40px;
-      height: 32px;
-    }
-  }
-}
-
-.chatMessage {
-  display: flex;
-  gap: 12px;
-  align-items: center;
-
-  .bubble {
-    position: relative;
-    background: rgba(128, 128, 128, 0.229);
-    border-radius: 4px;
-    padding: 4px 12px;
-
-    &:after {
-      content: "";
-      position: absolute;
-      left: 0;
-      top: 50%;
-      width: 0;
-      height: 0;
-      border: 10px solid transparent;
-      border-right-color: rgba(128, 128, 128, 0.229);
-      border-left: 0;
-      border-bottom: 0;
-      margin-top: -5px;
-      margin-left: -10px;
-    }
-  }
-}
-
-.chatMessage_isMe {
-  justify-content: end;
-
-  img {
-    display: none;
-  }
-
-  .bubble:after {
-    border-right-color: transparent;
-  }
-}
-
-.profile {
-  display: flex;
-  gap: 8px;
-
-  &_image {
-    border-radius: 50%;
-  }
-}
