+++ src/constant/index.ts
@@ -1,0 +1,7 @@
+export const FIRE_BASE_AUTH_ERROR = {
+  "auth/user-not-found": "이메일 혹은 비밀번호가 일치하지 않습니다.",
+  "auth/wrong-password": "이메일 혹은 비밀번호가 일치하지 않습니다.",
+  "auth/email-already-in-use": "이미 사용 중인 이메일입니다.",
+  "auth/invalid-email": "auth/invalid-email",
+  default: "로그인에 실패하였습니다.",
+};
