+++ src/firebase/firebase.ts
@@ -3,8 +3,6 @@ import { initializeApp } from "firebase/app";
 import { getAuth } from "firebase/auth";
 import { getDatabase } from "firebase/database";
 import { getFirestore } from "firebase/firestore";
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
 
 // Your web app's Firebase configuration
 const firebaseConfig = {
@@ -15,13 +13,11 @@ const firebaseConfig = {
   messagingSenderId: "802925921509",
   appId: "1:802925921509:web:042f2aa5ee62afd3f7000c",
   databaseURL: "https://homework-474df-default-rtdb.firebaseio.com",
-  
 };
 
 // Initialize Firebase
 const app = initializeApp(firebaseConfig);
 
-
 export const auth = getAuth(app);
-export const db = getFirestore(app)
+export const db = getFirestore(app);
 export const database = getDatabase(app);
