--- src/redux/slices/counterSlice.ts
@@ -1,59 +1,0 @@
-import { createAsyncThunk, createSlice, PayloadAction } from "@reduxjs/toolkit";
-import { RootState, AppThunk } from "./../store";
-// import { fetchCount } from "./counterAPI";
-
-export interface CounterState {
-  value: number;
-  status: "idle" | "loading" | "failed";
-}
-
-const initialState: CounterState = {
-  value: 0,
-  status: "idle",
-};
-
-// export const incrementAsync = createAsyncThunk(
-//   "counter/fetchCount",
-//   async (amount: number) => {
-//     const response = await fetchCount(amount);
-//     // The value we return becomes the `fulfilled` action payload
-//     return response.data;
-//   }
-// );
-
-export const counterSlice = createSlice({
-  name: "counter",
-  initialState,
-  // The `reducers` field lets us define reducers and generate associated actions
-  reducers: {
-    increment: (state) => {
-      state.value += 1;
-    },
-    decrement: (state) => {
-      state.value -= 1;
-    },
-    incrementByAmount: (state, action: PayloadAction<number>) => {
-      state.value += action.payload;
-    },
-  },
-
-  // extraReducers: (builder) => {
-  //   builder
-  //     .addCase(incrementAsync.pending, (state) => {
-  //       state.status = "loading";
-  //     })
-  //     .addCase(incrementAsync.fulfilled, (state, action) => {
-  //       state.status = "idle";
-  //       state.value += action.payload;
-  //     })
-  //     .addCase(incrementAsync.rejected, (state) => {
-  //       state.status = "failed";
-  //     });
-  // },
-});
-
-export const { increment, decrement, incrementByAmount } = counterSlice.actions;
-
-export const selectCount = (state: RootState) => state.counter.value;
-
-export default counterSlice.reducer;
