+++ todo.md
@@ -11,3 +11,10 @@
 
 [V] 로그인 및 레지스터 클라이언트 폼 없애기
 [] 에러가공 -constant
+
+### 12.10
+[V] 구글 로그인 말고 이름 -> 이메일 , 디폴트이미지
+[V] 첫번째 메시지 안보임 -> 저장이 안된줄 알았는데, 헤더에 가려짐
+[V] 채팅 주소 아이디 변경 - 이메일 조인
+[V] 상대편 이미지 디폴트에서 변경
+[V] 컴포넌트분리 - profile, chatMessage
